+++ src/main/java/insa/aubin/devisbatiment/Porte.java
@@ -1,8 +1,8 @@
 package insa.aubin.devisbatiment;
 
 public class Porte extends Ouverture {
-    private static final float LARGEUR_PORTE = 0.90f;
-    private static final float HAUTEUR_PORTE = 2.10f;
+    private static final double LARGEUR_PORTE = 0.90f;
+    private static final double HAUTEUR_PORTE = 2.10f;
 
     public Porte(String idPorte){
         super(idPorte, LARGEUR_PORTE, HAUTEUR_PORTE);
